+++ 2019/src/main/java/nl/remcoder/adventofcode/Day5.java
@@ -7,23 +7,23 @@ import java.util.concurrent.LinkedBlockingQueue;
 import java.util.stream.Stream;
 
 public class Day5 {
-    public int handlePart1(Stream<String> input) throws Exception {
+    public long handlePart1(Stream<String> input) throws Exception {
         String line = input.findFirst().orElseThrow(AssertionError::new);
 
-        int[] opcodes = Arrays.stream(line.split(","))
-                              .mapToInt(Integer::parseInt)
-                              .toArray();
+        long[] opcodes = Arrays.stream(line.split(","))
+                               .mapToLong(Long::parseLong)
+                               .toArray();
 
-        BlockingQueue<Integer> inputState = new LinkedBlockingQueue<>();
-        BlockingQueue<Integer> outputState = new LinkedBlockingQueue<>();
+        BlockingQueue<Long> inputState = new LinkedBlockingQueue<>();
+        BlockingQueue<Long> outputState = new LinkedBlockingQueue<>();
 
-        inputState.put(1);
+        inputState.put(1L);
 
         IntCodeComputer intCodeComputer = new IntCodeComputer(opcodes, inputState, outputState);
 
         intCodeComputer.runProgram();
 
-        int output = 0;
+        long output = 0;
 
         while (!outputState.isEmpty()) {
             output = outputState.take();
@@ -32,23 +32,23 @@ public class Day5 {
         return output;
     }
 
-    public int handlePart2(Stream<String> input) throws Exception {
+    public long handlePart2(Stream<String> input) throws Exception {
         String line = input.findFirst().orElseThrow(AssertionError::new);
 
-        int[] opcodes = Arrays.stream(line.split(","))
-                              .mapToInt(Integer::parseInt)
+        long[] opcodes = Arrays.stream(line.split(","))
+                              .mapToLong(Long::parseLong)
                               .toArray();
 
-        BlockingQueue<Integer> inputState = new LinkedBlockingQueue<>();
-        BlockingQueue<Integer> outputState = new LinkedBlockingQueue<>();
+        BlockingQueue<Long> inputState = new LinkedBlockingQueue<>();
+        BlockingQueue<Long> outputState = new LinkedBlockingQueue<>();
 
-        inputState.put(5);
+        inputState.put(5L);
 
         IntCodeComputer intCodeComputer = new IntCodeComputer(opcodes, inputState, outputState);
 
         intCodeComputer.runProgram();
 
-        int output = 0;
+        long output = 0;
 
         while (!outputState.isEmpty()) {
             output = outputState.take();
